After little changes
This project aims to dissect and understand the supply chain mechanism of Walmart, a global retail giant. By leveraging data visualization and analysis techniques, we will identify potential areas of improvement within the supply chain. This exploration will extend from procurement to the final delivery, with a particular focus on Logistical efficiency and cost-effectiveness.